requirements.txt
# Web Scraping
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0

# Browser Automation
selenium>=4.15.0
webdriver-manager>=4.0.0

# Data Processing
pandas>=2.0.0

# Utilities
fake-useragent>=1.4.0
python-dotenv>=1.0.0